pyproject.toml
[tool.poetry]
name = "ohm-fork-data"
version = "0.1.0"
description = ""
authors = ["Concave Community"]

[tool.poetry.dependencies]
python = "^3.8"
jupyter = "^1.0.0"
web3 = "^5.25.0"
requests = "^2.26.0"
PyYAML = "^6.0"
pandas = "^1.3.5"
ipython = "^7.30.1"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
